requirement.txt
asgiref==3.7.2
Django==4.1.13
django-cors-headers==4.3.1
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.6.1
Markdown==3.6
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.2.4
tzdata==2024.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
filelock==3.9.0
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
protobuf==4.24.3
requests==2.28.1
scipy==1.11.3
sympy==1.12
tensorboardX==2.6.2.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==3.0.0